# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Compra/GetComprasPorId": {
      "get": {
        "tags": [
          "Compra"
        ],
        "operationId": "GetComprasPorId",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Compra/CrearCompra": {
      "post": {
        "tags": [
          "Compra"
        ],
        "operationId": "CrearCompra",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompraForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompraDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompraDetailDTO": {
        "required": [
          "apellidoCliente",
          "direccionEnvio",
          "fechaCompra",
          "nombreCliente",
          "precioTotal",
          "tipoMetodoPago"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nombreCliente": {
            "minLength": 1,
            "type": "string",
            "format": "string"
          },
          "apellidoCliente": {
            "minLength": 1,
            "type": "string",
            "format": "string"
          },
          "direccionEnvio": {
            "minLength": 1,
            "type": "string",
            "format": "string"
          },
          "fechaCompra": {
            "type": "string",
            "format": "date"
          },
          "precioTotal": {
            "type": "number",
            "format": "currency",
            "readOnly": true
          },
          "tipoMetodoPago": {
            "$ref": "#/components/schemas/TiposMetodoPago"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraForCreateDTO": {
        "required": [
          "apellidoCliente",
          "direccionEnvio",
          "nombreCliente",
          "precioTotal",
          "tipoMetodoPago"
        ],
        "type": "object",
        "properties": {
          "nombreCliente": {
            "minLength": 1,
            "type": "string",
            "format": "string"
          },
          "apellidoCliente": {
            "minLength": 1,
            "type": "string",
            "format": "string"
          },
          "direccionEnvio": {
            "minLength": 1,
            "type": "string",
            "format": "string"
          },
          "precioTotal": {
            "type": "number",
            "format": "currency",
            "readOnly": true
          },
          "tipoMetodoPago": {
            "$ref": "#/components/schemas/TiposMetodoPago"
          },
          "compraItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompraItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CompraItemDTO": {
        "required": [
          "cantidad",
          "materialHerramienta",
          "nombreHerramienta",
          "precioHerramienta"
        ],
        "type": "object",
        "properties": {
          "HerramientaId": {
            "type": "integer",
            "format": "int32"
          },
          "nombreHerramienta": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "materialHerramienta": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "precioHerramienta": {
            "type": "number",
            "format": "currency"
          },
          "descripcionHerramienta": {
            "type": "string",
            "nullable": true
          },
          "cantidad": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "TiposMetodoPago": {
        "enum": [
          "TarjetaCredito",
          "PayPal",
          "Efectivo"
        ],
        "type": "string"
      },
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}